# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OneAquaHealth Insight API",
    "version": "1.0.0",
    "description": "Read-only API over the frozen OneAquaHealth snapshot (pipeline Stage 8). Every number was computed offline by the pipeline; the API only filters and reshapes. W (water quality, higher = cleaner) and H (contamination hazard, higher = riskier) are relative 0-100 indices anchored to citizen rubrics, not lab measurements, and always travel with a 90% interval (q05-q95) and an evidence flag. Based on volunteer observations, not laboratory tests. Not an official health advisory."
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
...
    "/api/v1/sites": {
      "get": {
        "tags": ["sites"],
        "summary": "Every site at the latest week, head to mouth",
        "description": "p_change and alert_level follow the sensitivity toggle. status is the Stage 7 status text, written for the primary model only (null under other variants).",
        "parameters": [{ "$ref": "#/components/parameters/Variant" }, { "$ref": "#/components/parameters/Sensitivity" }],
        "responses": {
          "200": { "$ref": "#/components/responses/SiteList" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/sites/{siteId}": {
      "get": {
        "tags": ["sites"],
        "summary": "One site: latest scores, neighbours, related findings and incidents",
        "parameters": [
          { "$ref": "#/components/parameters/SiteId" },
          { "$ref": "#/components/parameters/Variant" },
          { "$ref": "#/components/parameters/Sensitivity" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SiteDetail" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/sites/{siteId}/timeseries": {
      "get": {
        "tags": ["sites"],
        "summary": "Weekly W/H intervals, alerts and rainfall for one site",
        "parameters": [
          { "$ref": "#/components/parameters/SiteId" },
          { "$ref": "#/components/parameters/Variant" },
          { "$ref": "#/components/parameters/Sensitivity" },
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Timeseries" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/sites/{siteId}/observations": {
      "get": {
        "tags": ["sites"],
        "summary": "De-identified citizen reports for one site, newest first",
        "parameters": [
          { "$ref": "#/components/parameters/SiteId" },
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 100 } },
          { "$ref": "#/components/parameters/Offset" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Observations" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/scores": {
      "get": {
        "tags": ["network"],
        "summary": "Every site for one week (map / time slider)",
        "parameters": [
          { "$ref": "#/components/parameters/Variant" },
          { "$ref": "#/components/parameters/Sensitivity" },
          {
            "name": "week",
            "in": "query",
            "description": "Any date inside the grid; it maps to the week (Monday) that contains it. Defaults to the latest week.",
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/WeekScores" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/weeks": {
      "get": {
        "tags": ["network"],
        "summary": "Weekly rainfall (identical across sites) and rain-exposed flag",
        "parameters": [{ "$ref": "#/components/parameters/From" }, { "$ref": "#/components/parameters/To" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Weeks" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/river": {
      "get": {
        "tags": ["network"],
        "summary": "River graph: sites as nodes, directed upstream -> downstream edges",
        "responses": {
          "200": { "$ref": "#/components/responses/River" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/rain-pattern": {
      "get": {
        "tags": ["network"],
        "summary": "Rain -> overflow test (from the no-rain model, so the answer is not built in)",
        "responses": {
          "200": { "$ref": "#/components/responses/RainPattern" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/findings": {
      "get": {
        "tags": ["findings"],
        "summary": "Ranked plain-language findings",
        "description": "Written for the primary model at normal sensitivity. Every number in a finding's text appears in its facts object.",
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["incident", "site_status", "rain_pattern", "coverage"] } },
          { "name": "priority", "in": "query", "schema": { "type": "string", "enum": ["high", "medium", "low", "info"] } },
          { "name": "recent", "in": "query", "schema": { "type": "boolean" } },
          {
            "name": "site",
            "in": "query",
            "description": "The site's status finding plus findings for incidents that affected it",
            "schema": { "type": "string" }
          },
          { "name": "lang", "in": "query", "schema": { "type": "string", "default": "en" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 100 } },
          { "$ref": "#/components/parameters/Offset" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/FindingList" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/findings/{findingId}": {
      "get": {
        "tags": ["findings"],
        "summary": "One finding",
        "parameters": [
          {
            "name": "findingId",
            "in": "path",
            "required": true,
            "description": "e.g. incident:inc-015, site:gow-01, network:rain, network:coverage",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/FindingOne" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/incidents": {
      "get": {
        "tags": ["incidents"],
        "summary": "Incidents, most recent first",
        "parameters": [
          { "$ref": "#/components/parameters/Variant" },
          { "name": "recent", "in": "query", "description": "Ended within the last RECENT_WEEKS grid weeks", "schema": { "type": "boolean" } },
          { "name": "site", "in": "query", "description": "Only incidents that affected this site", "schema": { "type": "string" } },
          { "name": "level", "in": "query", "schema": { "type": "string", "enum": ["confirmed", "possible"] } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/IncidentList" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/incidents/{incidentId}": {
      "get": {
        "tags": ["incidents"],
        "summary": "One incident with its site-week alerts and its finding",
        "parameters": [
          { "name": "incidentId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/Variant" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/IncidentOne" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "Variant": {
        "name": "variant",
        "in": "query",
        "description": "Rain-assumption toggle: M1 = primary model with rain (default), M1_norain = model without rain. Allowed values are listed in /api/v1/meta.",
        "schema": { "type": "string", "enum": ["M1", "M1_norain"], "default": "M1" }
      },
      "Sensitivity": {
        "name": "sensitivity",
        "in": "query",
        "description": "Detection sensitivity: size of W drop counted as a change (sensitive 5, normal 10, strict 20 index points).",
        "schema": { "type": "string", "enum": ["sensitive", "normal", "strict"], "default": "normal" }
      },
      "From": { "name": "from", "in": "query", "description": "First week_start to include (default: first grid week)", "schema": { "type": "string", "format": "date" } },
      "To": { "name": "to", "in": "query", "description": "Last week_start to include (default: latest grid week)", "schema": { "type": "string", "format": "date" } },
      "Offset": { "name": "offset", "in": "query", "schema": { "type": "integer", "minimum": 0, "default": 0 } },
      "SiteId": { "name": "siteId", "in": "path", "required": true, "description": "e.g. gow-01 (head) to gow-10 (mouth)", "schema": { "type": "string" } }
    },
    "responses": {
      "BadRequest": { "description": "Invalid, unknown or repeated parameter", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Unknown id or route", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NoSnapshot": { "description": "No snapshot loaded yet", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "SiteList": {
        "description": "Sites at the latest week",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/SiteLatest" } } } }] } } }
      },
      "SiteDetail": {
        "description": "One site",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "properties": { "data": { "$ref": "#/components/schemas/SiteDetail" } } }] } } }
      },
      "Timeseries": {
        "description": "Weekly series",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "properties": { "site": { "$ref": "#/components/schemas/SiteRef" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/TimeseriesWeek" } } } }] } } }
      },
      "Observations": {
        "description": "Reports",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "properties": { "site": { "$ref": "#/components/schemas/SiteRef" }, "total": { "type": "integer" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Observation" } } } }] } } }
      },
      "WeekScores": {
        "description": "Every site in one week",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "properties": { "week": { "$ref": "#/components/schemas/Week" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/WeekScore" } } } }] } } }
      },
      "Weeks": {
        "description": "Weeks",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Week" } } } }] } } }
      },
      "River": {
        "description": "River graph",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "properties": { "data": { "$ref": "#/components/schemas/River" } } }] } } }
      },
      "RainPattern": {
        "description": "Rain -> overflow result",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "properties": { "data": { "$ref": "#/components/schemas/RainPattern" } } }] } } }
      },
      "FindingList": {
        "description": "Findings",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "properties": { "total": { "type": "integer" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Finding" } } } }] } } }
      },
      "FindingOne": {
        "description": "Finding",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "properties": { "data": { "$ref": "#/components/schemas/Finding" } } }] } } }
      },
      "IncidentList": {
        "description": "Incidents",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/ListEnvelope" }, { "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Incident" } } } }] } } }
      },
      "IncidentOne": {
        "description": "Incident with alerts",
        "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "properties": { "data": { "$ref": "#/components/schemas/IncidentDetail" } } }] } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["status", "message"],
            "properties": { "status": { "type": "integer" }, "message": { "type": "string" }, "details": { "type": "object" } }
          }
        }
      },
...
      "Envelope": {
        "type": "object",
        "required": ["snapshot", "disclaimer", "data"],
        "properties": {
          "snapshot": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "description": "Content address of the snapshot; also sent as the X-OAH-Snapshot header" },
              "schema_version": { "type": "string" },
              "latest_week": { "type": "string", "format": "date" },
              "created_at_utc": { "type": "string" }
            }
          },
          "disclaimer": { "type": "string" },
          "params": { "type": "object", "description": "The parameters used, after defaults" },
          "note": { "type": "string", "description": "What the toggles do and do not change for this endpoint" },
          "data": {}
        }
      },
      "ListEnvelope": {
        "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "properties": { "count": { "type": "integer" }, "data": { "type": "array" } } }]
      },
...
      "SiteRef": { "type": "object", "properties": { "site_id": { "type": "string" }, "name": { "type": "string" } } },
...
      "SiteLatest": {
        "type": "object",
        "properties": {
          "site_id": { "type": "string" }, "name": { "type": "string" }, "display_order": { "type": "integer" },
          "lat": { "type": "number" }, "lon": { "type": "number" }, "watershed": { "type": "string" },
          "distance_from_mouth_m": { "type": "number" }, "is_cso_outfall_adjacent": { "type": "boolean" },
          "known_context_note": { "type": ["string", "null"] }, "exposure_weight": { "type": ["number", "null"] },
          "week_start": { "type": "string", "format": "date" },
          "W": { "$ref": "#/components/schemas/Band" }, "H": { "$ref": "#/components/schemas/Band" },
          "evidence": { "$ref": "#/components/schemas/Evidence" },
          "mixing_flag": { "type": "boolean", "description": "Bimodal posterior: treat the interval with care" },
          "n_reports": { "type": "integer" },
          "p_change": { "type": "number", "description": "P(W dropped by more than the sensitivity threshold)" },
          "alert_level": { "$ref": "#/components/schemas/AlertLevel" },
          "status": { "type": ["object", "null"], "properties": { "headline": { "type": "string" }, "body": { "type": "string" } } },
          "exposure": { "type": "array", "items": { "$ref": "#/components/schemas/Exposure" } }
        }
      },
      "SiteDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/SiteLatest" },
          {
            "properties": {
              "upstream": { "type": "array", "items": { "$ref": "#/components/schemas/Neighbour" } },
              "downstream": { "type": "array", "items": { "$ref": "#/components/schemas/Neighbour" } },
              "findings": { "type": "array", "items": { "$ref": "#/components/schemas/Finding" } },
              "incidents": { "type": "array", "items": { "$ref": "#/components/schemas/Incident" } }
            }
          }
        ]
      },
...
      "TimeseriesWeek": {
        "type": "object",
        "properties": {
          "week_start": { "type": "string", "format": "date" },
          "W": { "$ref": "#/components/schemas/Band" }, "H": { "$ref": "#/components/schemas/Band" },
          "evidence": { "$ref": "#/components/schemas/Evidence" }, "mixing_flag": { "type": "boolean" }, "n_reports": { "type": "integer" },
          "p_change": { "type": "number" }, "alert_level": { "$ref": "#/components/schemas/AlertLevel" },
          "episode_id": { "type": ["string", "null"] },
          "weekly_rainfall_mm": { "type": "number" }, "heavy_rain_week": { "type": "boolean" }, "rain_exposed": { "type": "boolean" }
        }
      },
...
      "Observation": {
        "type": "object",
        "description": "De-identified: no observer id, raw payload or time of day",
        "properties": {
          "observation_id": { "type": "string" }, "observed_date": { "type": "string", "format": "date" },
          "week_start": { "type": "string", "format": "date" },
          "water_clarity": { "type": ["number", "null"], "description": "1 = clear, 5 = murky" },
          "smell": { "type": ["string", "null"] }, "smell_intensity": { "type": ["number", "null"], "description": "0-3" },
          "trash_level": { "type": ["number", "null"] }, "insect_presence": { "type": ["boolean", "null"] },
          "insect_diversity": { "type": ["number", "null"] }, "notes": { "type": ["string", "null"] },
          "source": { "type": "string", "enum": ["simulated", "real"] }
        }
      },
...
    }
  }
}